added balanced metadata
